refactor: use updated hooks and utilities in stories
